+++ client/capi/soem_link/soem_link.h
@@ -3,7 +3,7 @@
 // Created Date: 20/02/2021
 // Author: Shun Suzuki
 // -----
-// Last Modified: 19/07/2021
+// Last Modified: 27/09/2021
 // Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
 // -----
 // Copyright (c) 2021 Hapis Lab. All rights reserved.
@@ -15,11 +15,11 @@
 
 #include "../base/header.h"
 
-using ErrorHandler = void (*)(const char*);
+using OnLostCallback = void (*)(const char*);
 
 extern "C" {
 EXPORT_AUTD int32_t AUTDGetAdapterPointer(void** out);
 EXPORT_AUTD void AUTDGetAdapter(void* p_adapter, int32_t index, char* desc, char* name);
 EXPORT_AUTD void AUTDFreeAdapterPointer(void* p_adapter);
-EXPORT_AUTD void AUTDLinkSOEM(void** out, const char* ifname, int32_t device_num, uint32_t cycle_ticks, ErrorHandler handler);
+EXPORT_AUTD void AUTDLinkSOEM(void** out, const char* ifname, int32_t device_num, uint32_t cycle_ticks, OnLostCallback callback);
 }