+++ client/examples/examples/advanced.hpp
@@ -3,7 +3,7 @@
 // Created Date: 19/05/2021
 // Author: Shun Suzuki
 // -----
-// Last Modified: 13/08/2021
+// Last Modified: 22/09/2021
 // Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
 // -----
 // Copyright (c) 2021 Hapis Lab. All rights reserved.
@@ -22,8 +22,8 @@ class BurstModulation final : public autd::core::Modulation {
     return std::make_shared<BurstModulation>(buf_size, mod_freq_div);
   }
   void calc() override {
-    this->buffer().resize(_buf_size, 0);
-    this->buffer().at(_buf_size - 1) = 0xFF;
+    this->_buffer.resize(_buf_size, 0);
+    this->_buffer[_buf_size - 1] = 0xFF;
   }
 
   BurstModulation(const size_t buf_size, const uint16_t mod_freq_div) : Modulation(mod_freq_div), _buf_size(buf_size) {}
@@ -38,7 +38,6 @@ class UniformGain final : public autd::core::Gain {
   void calc(const autd::GeometryPtr& geometry) override {
     for (size_t i = 0; i < geometry->num_transducers(); i++)
       this->_data[geometry->device_idx_for_trans_idx(i)].at(i % autd::NUM_TRANS_IN_UNIT) = 0xFF80;
-    this->_built = true;
   }
 };