+++ client/include/autd3/core/firmware_version.hpp
@@ -3,7 +3,7 @@
 // Created Date: 14/04/2021
 // Author: Shun Suzuki
 // -----
-// Last Modified: 26/07/2021
+// Last Modified: 28/09/2021
 // Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
 // -----
 // Copyright (c) 2021 Hapis Lab. All rights reserved.
@@ -49,16 +49,12 @@ class FirmwareInfo {
       ss << "v0." << version_number + 3;
       return ss.str();
     }
-    if (0x000A <= version_number && version_number <= 0x0011) {
+    if (0x000A <= version_number && version_number <= 0x0012) {
       ss << "v1." << version_number - 0x000A;
       return ss.str();
     }
     if (version_number == 0xFFFF) {
       return "emulator";
-    }
-    if ((version_number & 0xF000) == 0x1000) {
-      ss << "v" << version_number - 0x1000 + 1 << "-lite";
-      return ss.str();
     }
     if ((version_number & 0xF000) == 0xF000) {
       ss << "v0." << version_number - 0xF000 + 1 << "-freq-shift";