+++ client/lib/gain/primitive.cpp
@@ -3,7 +3,7 @@
 // Created Date: 14/04/2021
 // Author: Shun Suzuki
 // -----
-// Last Modified: 03/09/2021
+// Last Modified: 22/09/2021
 // Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
 // -----
 // Copyright (c) 2021 Hapis Lab. All rights reserved.
@@ -34,7 +34,6 @@ void Grouped::calc(const core::GeometryPtr& geometry) {
     auto group_id = geometry->group_id_for_device_idx(i);
     this->_data[i] = _gain_map.count(group_id) ? _gain_map[group_id]->data()[i] : DataArray{0x0000};
   }
-  this->_built = true;
 }
 
 GainPtr PlaneWave::create(const Vector3& direction, const double amp) { return create(direction, core::Utilities::to_duty(amp)); }
@@ -44,57 +43,52 @@ GainPtr PlaneWave::create(const Vector3& direction, uint8_t duty) { return std::
 void PlaneWave::calc(const core::GeometryPtr& geometry) {
   const auto dir = this->_direction.normalized();
 
-  const auto ultrasound_wavelength = geometry->wavelength();
+  const auto wavenum = 2.0 * M_PI / geometry->wavelength();
   for (size_t dev = 0; dev < geometry->num_devices(); dev++)
     for (size_t i = 0; i < NUM_TRANS_IN_UNIT; i++) {
       const auto dist = geometry->position(dev, i).dot(dir);
-      const auto phase = core::Utilities::to_phase(dist / ultrasound_wavelength);
+      const auto phase = core::Utilities::to_phase(dist * wavenum);
       this->_data[dev][i] = core::Utilities::pack_to_u16(this->_duty, phase);
     }
-
-  this->_built = true;
 }
 
 GainPtr FocalPoint::create(const Vector3& point, const double amp) { return create(point, core::Utilities::to_duty(amp)); }
 GainPtr FocalPoint::create(const Vector3& point, uint8_t duty) { return std::make_shared<FocalPoint>(point, duty); }
 
 void FocalPoint::calc(const core::GeometryPtr& geometry) {
-  const auto ultrasound_wavelength = geometry->wavelength();
+  const auto wavenum = 2.0 * M_PI / geometry->wavelength();
   for (size_t dev = 0; dev < geometry->num_devices(); dev++)
     for (size_t i = 0; i < NUM_TRANS_IN_UNIT; i++) {
       const auto dist = (geometry->position(dev, i) - this->_point).norm();
-      const auto phase = core::Utilities::to_phase(dist / ultrasound_wavelength);
+      const auto phase = core::Utilities::to_phase(dist * wavenum);
       this->_data[dev][i] = core::Utilities::pack_to_u16(this->_duty, phase);
     }
-
-  this->_built = true;
 }
 
-GainPtr BesselBeam::create(const Vector3& point, const Vector3& vec_n, const double theta_z, const double amp) {
-  return create(point, vec_n, theta_z, core::Utilities::to_duty(amp));
+GainPtr BesselBeam::create(const Vector3& apex, const Vector3& vec_n, const double theta_z, const double amp) {
+  return create(apex, vec_n, theta_z, core::Utilities::to_duty(amp));
 }
 
-GainPtr BesselBeam::create(const Vector3& point, const Vector3& vec_n, double theta_z, uint8_t duty) {
-  return std::make_shared<BesselBeam>(point, vec_n, theta_z, duty);
+GainPtr BesselBeam::create(const Vector3& apex, const Vector3& vec_n, double theta_z, uint8_t duty) {
+  return std::make_shared<BesselBeam>(apex, vec_n, theta_z, duty);
 }
 
 void BesselBeam::calc(const core::GeometryPtr& geometry) {
-  if (_vec_n.norm() > 0) _vec_n = _vec_n.normalized();
-  const Vector3 v(_vec_n.y(), -_vec_n.x(), 0.);
+  _vec_n.normalize();
+  Vector3 v = Vector3::UnitZ().cross(_vec_n);
+  const auto theta_v = std::asin(v.norm());
+  v.normalize();
+  const Eigen::AngleAxisd rot(-theta_v, v);
 
-  const auto theta_w = std::asin(v.norm());
-
-  const auto ultrasound_wavelength = geometry->wavelength();
+  const auto wavenum = 2.0 * M_PI / geometry->wavelength();
   for (size_t dev = 0; dev < geometry->num_devices(); dev++)
     for (size_t i = 0; i < NUM_TRANS_IN_UNIT; i++) {
-      const auto r = geometry->position(dev, i) - this->_point;
-      const auto v_x_r = r.cross(v);
-      const auto rr = std::cos(theta_w) * r + std::sin(theta_w) * v_x_r + v.dot(r) * (1.0 - std::cos(theta_w)) * v;
+      const auto r = geometry->position(dev, i) - this->_apex;
+      const auto rr = rot * r;
       const auto d = std::sin(_theta_z) * std::sqrt(rr.x() * rr.x() + rr.y() * rr.y()) - std::cos(_theta_z) * rr.z();
-      const auto phase = core::Utilities::to_phase(d / ultrasound_wavelength);
+      const auto phase = core::Utilities::to_phase(d * wavenum);
       this->_data[dev][i] = core::Utilities::pack_to_u16(this->_duty, phase);
     }
-  this->_built = true;
 }
 
 GainPtr Custom::create(const uint16_t* data, const size_t data_length) {
@@ -109,10 +103,7 @@ GainPtr Custom::create(const uint16_t* data, const size_t data_length) {
 
 GainPtr Custom::create(const std::vector<DataArray>& data) { return std::make_shared<Custom>(data); }
 
-void Custom::calc(const core::GeometryPtr&) {
-  this->_data = std::move(this->_raw_data);
-  this->_built = true;
-}
+void Custom::calc(const core::GeometryPtr&) { this->_data = std::move(this->_raw_data); }
 
 GainPtr TransducerTest::create(const size_t transducer_index, const uint8_t duty, const uint8_t phase) {
   return std::make_shared<TransducerTest>(transducer_index, duty, phase);
@@ -121,7 +112,5 @@ GainPtr TransducerTest::create(const size_t transducer_index, const uint8_t duty
 void TransducerTest::calc(const core::GeometryPtr& geometry) {
   this->_data[geometry->device_idx_for_trans_idx(_transducer_idx)][_transducer_idx % NUM_TRANS_IN_UNIT] =
       core::Utilities::pack_to_u16(this->_duty, this->_phase);
-
-  this->_built = true;
 }
 }  // namespace autd::gain