+++ client/lib/link/soem/autdsoem.hpp
@@ -3,7 +3,7 @@
 // Created Date: 08/03/2021
 // Author: Shun Suzuki
 // -----
-// Last Modified: 19/07/2021
+// Last Modified: 29/09/2021
 // Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
 // -----
 // Copyright (c) 2021 Hapis Lab. All rights reserved.
@@ -12,8 +12,10 @@
 #pragma once
 
 #include <atomic>
+#include <condition_variable>
 #include <functional>
 #include <memory>
+#include <mutex>
 #include <string>
 #include <utility>
 #include <vector>
@@ -29,15 +31,16 @@ struct SOEMCallback final : core::CallbackHandler {
   SOEMCallback(SOEMCallback&& obj) = delete;
   SOEMCallback& operator=(SOEMCallback&& obj) = delete;
 
-  explicit SOEMCallback(const int expected_wkc, std::function<bool()> error_handler)
-      : _autd3_rt_lock(false), _expected_wkc(expected_wkc), _error_handler(std::move(error_handler)) {}
+  explicit SOEMCallback(const int expected_wkc, std::function<bool()> error_handler, std::atomic<bool>* sent)
+      : _rt_lock(false), _expected_wkc(expected_wkc), _error_handler(std::move(error_handler)), _sent(sent) {}
 
   void callback() override;
 
  private:
-  std::atomic<bool> _autd3_rt_lock;
+  std::atomic<bool> _rt_lock;
   int _expected_wkc;
   std::function<bool()> _error_handler;
+  std::atomic<bool>* _sent;
 };
 
 struct ECConfig {
@@ -47,6 +50,8 @@ struct ECConfig {
   size_t body_size;
   size_t input_frame_size;
 };
+
+constexpr size_t SEND_BUF_SIZE = 32;
 
 class SOEMController {
  public:
@@ -58,19 +63,19 @@ class SOEMController {
   SOEMController& operator=(SOEMController&& obj) = delete;
 
   void open(const char* ifname, size_t dev_num, ECConfig config);
-  void set_lost_handler(std::function<void(std::string)> handler);
+  void on_lost(std::function<void(std::string)> callback);
   void close();
 
   [[nodiscard]] bool is_open() const;
 
-  void send(const uint8_t* buf, size_t size) const;
+  void send(const uint8_t* buf, size_t size);
   void read(uint8_t* rx) const;
 
  private:
   void setup_sync0(bool activate, uint32_t cycle_time_ns) const;
 
   bool error_handle();
-  std::function<void(std::string)> _link_lost_handle = nullptr;
+  std::function<void(std::string)> _on_lost = nullptr;
 
   std::unique_ptr<uint8_t[]> _io_map;
   size_t _io_map_size;
@@ -78,6 +83,14 @@ class SOEMController {
   size_t _dev_num;
   ECConfig _config;
   bool _is_open;
+
+  std::unique_ptr<std::unique_ptr<uint8_t[]>[]> _send_buf;
+  size_t _send_buf_cursor;
+  size_t _send_buf_size;
+  std::mutex _send_mtx;
+  std::condition_variable _send_cond;
+  std::thread _send_thread;
+  std::atomic<bool> _sent;
 
   std::unique_ptr<core::Timer<SOEMCallback>> _timer;
 };